Move pkg/internal/tecdsatest to pkg/tecdsa/internal/test
For consistency with changes proposed in https://github.com/keep-network/keep-core/pull/3626#discussion_r1227884983
we move the tecdsatest package as it's used internally in pkg/tecdsa only.
