+++ pkg/tecdsa/internal/test/testdata/private_key_share_data_0.json
